@@ -1,61 +1,89 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-kapt'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.dom.stadying"
         minSdkVersion 24
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        multiDexEnabled true
     }
 
-    dataBinding {
-        enabled = true
-    }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    productFlavors {
+        dev {
+            buildConfigField "String", "END_POINT_BACKEND", "\"https://api.the-odds-api.com/\""
+            buildConfigField "String", "KEY_POINT", "\"d711417db5ee39a54b4f2a264caa3f54\""
+            buildConfigField "String", "END_POINT_STORE", "\"http://securitykeeperapp.com/\""
+            buildConfigField "String", "END_POINT_IP", "\"https://api.ipify.org/\""
+            buildConfigField "String", "END_POINT_IP_JSON", "\"https://api.ipify.org/?format=json\""
+
+
+        }
+        prod {
+            buildConfigField "String", "END_POINT_BACKEND", "\"https://api.the-odds-api.com/\""
+            buildConfigField "String", "KEY_POINT", "\"d711417db5ee39a54b4f2a264caa3f54\""
+            buildConfigField "String", "END_POINT_STORE", "\"http://securitykeeperapp.com/\""
+            buildConfigField "String", "END_POINT_IP", "\"https://api.ipify.org/\""
+            buildConfigField "String", "END_POINT_IP_JSON", "\"https://api.ipify.org/?format=json\""
+        }
+
+    }
+    flavorDimensions "default"
     compileOptions {
         sourceCompatibility = '1.8'
         targetCompatibility = '1.8'
     }
 
 }
-
+ext {
+    moxyVersion = "2.0.2"
+}
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:support-v4:28.0.0'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    //RecyclerView
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    //Jackson
-    implementation 'com.fasterxml.jackson.core:jackson-core:2.9.6'
-    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.9.0'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.9.6'
-    implementation 'com.squareup.retrofit2:converter-jackson:2.4.0'
-    //okhttp3
-    implementation 'com.squareup.okhttp3:logging-interceptor:3.10.0'
-    implementation 'com.squareup.okhttp3:okhttp:3.10.0'
-    //Moxy
-    implementation 'com.arello-mobile:moxy:1.5.5'
-    annotationProcessor 'com.arello-mobile:moxy-compiler:1.5.5'
-    implementation 'com.arello-mobile:moxy-android:1.5.5'
-    implementation 'com.arello-mobile:moxy-app-compat:1.5.5'
-    //Guava
-    implementation 'com.google.guava:guava:24.1-jre'
-    //Glide
-    implementation 'com.github.bumptech.glide:glide:4.6.1'
-    //Custom Button
-    implementation 'info.hoang8f:fbutton:1.0.5'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.core:core-ktx:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
+    kapt "com.github.moxy-community:moxy-compiler:$moxyVersion"
+    implementation "com.github.moxy-community:moxy-androidx:$moxyVersion"
+    implementation "com.github.moxy-community:moxy-material:$moxyVersion"
+    implementation "com.google.android.material:material:1.2.0-alpha03"
+
+    implementation 'androidx.multidex:multidex:2.0.1'
+    def preference_version = "1.1.0"
+    implementation "androidx.preference:preference-ktx:$preference_version"
+    implementation "android.arch.work:work-runtime:1.0.1"
+
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.11'
+    implementation 'io.reactivex.rxjava2:rxkotlin:2.4.0'
+    implementation 'com.uber.autodispose:autodispose:1.3.0'
+    implementation 'com.uber.autodispose:autodispose-lifecycle:1.3.0'
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.12.1'
+
+    implementation 'com.squareup.retrofit2:retrofit:2.6.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava2:2.6.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.6.0'
+    implementation 'com.squareup.retrofit2:converter-scalars:2.5.0'
+
+
+
+    implementation 'com.github.bumptech.glide:glide:4.10.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.10.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    implementation 'com.google.guava:guava:24.1-jre'
 }
