@@ -1,6 +1,6 @@
 package com.alphilippov.studyingmap.databind;
 
-import android.databinding.BindingAdapter;
+import androidx.databinding.BindingAdapter;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
